Split units utilities from cata_utility
Hopefully should reduce the cost of including cata_utility.h and speed
up build times.
