+++ src/cata_utility.h
@@ -8,13 +8,12 @@
 #include <utility>
 #include <vector>
 #include <algorithm>
+#include <memory>
 #include <type_traits>
 
 #if defined (_WIN32) && !defined (_MSC_VER)
 #include <ext/stdio_filebuf.h>
 #endif
-
-#include "units.h"
 
 class JsonIn;
 class JsonOut;
@@ -38,15 +37,6 @@ struct pair_greater_cmp_first {
 struct null_deleter {
     template<typename T>
     void operator()( T * ) const {}
-};
-
-/**
- * Type of object that a measurement is taken on.  Used, for example, to display wind speed in m/s
- * while displaying vehicle speed in km/h.
- */
-enum units_type {
-    VU_VEHICLE,
-    VU_WIND
 };
 
 /**
@@ -76,16 +66,6 @@ template<typename T, typename std::enable_if<std::is_integral<T>::value, int>::t
 T divide_round_up( T num, T den )
 {
     return ( num + den - 1 ) / den;
-}
-
-/** Divide @p num by @p den, rounding up
- *
- * @p num must be non-negative, @p den must be positive, and @c num+den must not overflow.
- */
-template<typename T, typename U>
-T divide_round_up( units::quantity<T, U> num, units::quantity<T, U> den )
-{
-    return divide_round_up( num.value(), den.value() );
 }
 
 /**
@@ -175,80 +155,6 @@ double logarithmic_range( int min, int max, int pos );
  * @returns Value of mod, possibly altered to respect the min and max boundaries
  */
 int bound_mod_to_vals( int val, int mod, int max, int min );
-
-/**
- * Create a units label for a velocity value.
- *
- * Gives the name of the velocity unit in the user selected unit system, either
- * "km/h", "ms/s" or "mph".  Used to add abbreviated unit labels to the output of
- * @ref convert_velocity.
- *
- * @param vel_units type of velocity desired (i.e. wind or vehicle)
- *
- * @return name of unit.
- */
-const char *velocity_units( units_type vel_units );
-
-/**
- * Create a units label for a weight value.
- *
- * Gives the name of the weight unit in the user selected unit system, either
- * "kgs" or "lbs".  Used to add unit labels to the output of @ref convert_weight.
- *
- * @return name of unit
- */
-const char *weight_units();
-
-/**
- * Create an abbreviated units label for a volume value.
- *
- * Returns the abbreviated name for the volume unit for the user selected unit system,
- * i.e. "c", "L", or "qt". Used to add unit labels to the output of @ref convert_volume.
- *
- * @return name of unit.
- */
-const char *volume_units_abbr();
-
-/**
- * Create a units label for a volume value.
- *
- * Returns the abbreviated name for the volume unit for the user selected unit system,
- * ie "cup", "liter", or "quart". Used to add unit labels to the output of @ref convert_volume.
- *
- * @return name of unit.
- */
-const char *volume_units_long();
-
-/**
- * Convert internal velocity units to units defined by user.
- *
- * @param velocity A velocity value in internal units.
- * @param vel_units General type of item this velocity is for (e.g. vehicles or wind)
- *
- * @returns Velocity in the user selected measurement system and in appropriate
- *          units for the object being measured.
- */
-double convert_velocity( int velocity, units_type vel_units );
-
-/**
- * Convert weight in grams to units defined by user (kg or lbs)
- *
- * @param weight to be converted.
- *
- * @returns Weight converted to user selected unit
- */
-double convert_weight( const units::mass &weight );
-
-/**
- * Convert volume from ml to units defined by user.
- */
-double convert_volume( int volume );
-
-/**
- * Convert volume from ml to units defined by user,
- * optionally returning the units preferred scale.
- */
-double convert_volume( int volume, int *out_scale );
 
 /**
  * Clamp (number and space wise) value to with,
