+++ src/iexamine.cpp
@@ -88,6 +88,7 @@
 #include "ui_manager.h"
 #include "uistate.h"
 #include "units.h"
+#include "units_utility.h"
 #include "value_ptr.h"
 #include "vpart_position.h"
 #include "weather.h"
