+++ src/units_utility.cpp
@@ -1,0 +1,111 @@
+#include "units_utility.h"
+
+#include "options.h"
+
+const char *velocity_units( const units_type vel_units )
+{
+    if( get_option<std::string>( "USE_METRIC_SPEEDS" ) == "mph" ) {
+        return _( "mph" );
+    } else if( get_option<std::string>( "USE_METRIC_SPEEDS" ) == "t/t" ) {
+        //~ vehicle speed tiles per turn
+        return _( "t/t" );
+    } else {
+        switch( vel_units ) {
+            case VU_VEHICLE:
+                return _( "km/h" );
+            case VU_WIND:
+                return _( "m/s" );
+        }
+    }
+    return "error: unknown units!";
+}
+
+const char *weight_units()
+{
+    return get_option<std::string>( "USE_METRIC_WEIGHTS" ) == "lbs" ? _( "lbs" ) : _( "kg" );
+}
+
+const char *volume_units_abbr()
+{
+    const std::string vol_units = get_option<std::string>( "VOLUME_UNITS" );
+    if( vol_units == "c" ) {
+        return pgettext( "Volume unit", "c" );
+    } else if( vol_units == "l" ) {
+        return pgettext( "Volume unit", "L" );
+    } else {
+        return pgettext( "Volume unit", "qt" );
+    }
+}
+
+const char *volume_units_long()
+{
+    const std::string vol_units = get_option<std::string>( "VOLUME_UNITS" );
+    if( vol_units == "c" ) {
+        return _( "cup" );
+    } else if( vol_units == "l" ) {
+        return _( "liter" );
+    } else {
+        return _( "quart" );
+    }
+}
+
+double convert_velocity( int velocity, const units_type vel_units )
+{
+    const std::string type = get_option<std::string>( "USE_METRIC_SPEEDS" );
+    // internal units to mph conversion
+    double ret = static_cast<double>( velocity ) / 100;
+
+    if( type == "km/h" ) {
+        switch( vel_units ) {
+            case VU_VEHICLE:
+                // mph to km/h conversion
+                ret *= 1.609f;
+                break;
+            case VU_WIND:
+                // mph to m/s conversion
+                ret *= 0.447f;
+                break;
+        }
+    } else if( type == "t/t" ) {
+        ret /= 4;
+    }
+
+    return ret;
+}
+
+double convert_weight( const units::mass &weight )
+{
+    double ret = to_gram( weight );
+    if( get_option<std::string>( "USE_METRIC_WEIGHTS" ) == "kg" ) {
+        ret /= 1000;
+    } else {
+        ret /= 453.6;
+    }
+    return ret;
+}
+
+double convert_volume( int volume )
+{
+    return convert_volume( volume, nullptr );
+}
+
+double convert_volume( int volume, int *out_scale )
+{
+    double ret = volume;
+    int scale = 0;
+    const std::string vol_units = get_option<std::string>( "VOLUME_UNITS" );
+    if( vol_units == "c" ) {
+        ret *= 0.004;
+        scale = 1;
+    } else if( vol_units == "l" ) {
+        ret *= 0.001;
+        scale = 2;
+    } else {
+        ret *= 0.00105669;
+        scale = 2;
+    }
+    if( out_scale != nullptr ) {
+        *out_scale = scale;
+    }
+    return ret;
+}
